pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eclipse-ecor"
version = "0.1.0"
description = "Evolutionary Clinical Intake and Prioritization System - LLM-guided triage mechanism design"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "ECLIPSE Contributors"}
]
keywords = [
    "emergency department",
    "triage",
    "evolutionary algorithm",
    "healthcare operations",
    "simulation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
llm = ["openai>=1.55"]

[project.urls]
Repository = "https://github.com/your-org/ECLIPSE"
Documentation = "https://github.com/your-org/ECLIPSE#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["baselines*", "eval*", "evolution*", "mechanisms*", "sim*", "scripts*"]
